fix: address resume upload e2e review
