+++ .github/workflows/e2e-tests.yml
@@ -144,25 +144,11 @@ jobs:
         run: npx playwright install --with-deps chromium
 
       - name: Start MinIO
-        shell: bash
-        run: |
-          set -euo pipefail
-          docker run -d --rm \
-            --name factory-careers-security-minio \
-            -e MINIO_ROOT_USER=e2e-access-key \
-            -e MINIO_ROOT_PASSWORD=e2e-secret-key \
-            -p 127.0.0.1:9000:9000 \
-            minio/minio server /data
-
-          for _ in $(seq 1 30); do
-            if curl -fsS http://127.0.0.1:9000/minio/health/live >/dev/null; then
-              exit 0
-            fi
-            sleep 1
-          done
-
-          docker logs factory-careers-security-minio
-          exit 1
+        uses: ./.github/actions/start-minio
+        with:
+          container-name: factory-careers-security-minio
+          access-key: e2e-access-key
+          secret-key: e2e-secret-key
 
       - name: Run core tenant isolation security checks
         run: npm run test:e2e:security:core
@@ -236,25 +222,11 @@ jobs:
         run: npx playwright install --with-deps chromium
 
       - name: Start MinIO
-        shell: bash
-        run: |
-          set -euo pipefail
-          docker run -d --rm \
-            --name factory-careers-uploads-minio \
-            -e MINIO_ROOT_USER=e2e-access-key \
-            -e MINIO_ROOT_PASSWORD=e2e-secret-key \
-            -p 127.0.0.1:9000:9000 \
-            minio/minio server /data
-
-          for _ in $(seq 1 30); do
-            if curl -fsS http://127.0.0.1:9000/minio/health/live >/dev/null; then
-              exit 0
-            fi
-            sleep 1
-          done
-
-          docker logs factory-careers-uploads-minio
-          exit 1
+        uses: ./.github/actions/start-minio
+        with:
+          container-name: factory-careers-uploads-minio
+          access-key: e2e-access-key
+          secret-key: e2e-secret-key
 
       - name: Run resume upload browser checks
         run: npm run test:e2e:uploads
